fix: inherit dtype  from positions

+++ src/__init__.py
@@ -84,7 +84,7 @@ def interpolate(
     if grid_data.ndim == 3:
         grid_data = cp.ascontiguousarray(grid_data[:, :, :, cp.newaxis])
     nf = grid_data.shape[3]
-    result = cp.zeros((pos.shape[0], nf), dtype=cp.float32)
+    result = cp.zeros((pos.shape[0], nf), dtype=pos.dtype)
     interpolateField(pos, grid_data, result, L, gradient, gradient_direction, kernel)
     return result
 
@@ -134,6 +134,6 @@ def spread(
         quantity = cp.array(quantity)
     if kernel is None:
         kernel = create_kernel("peskin3pt")
-    result = cp.zeros((n[0], n[1], n[2], quantity.shape[-1]), dtype=cp.float32)
+    result = cp.zeros((n[0], n[1], n[2], quantity.shape[-1]), dtype=pos.dtype)
     spreadParticles(pos, quantity, result, L, n, gradient, gradient_direction, kernel)
     return result
